+++ src/vulkan_linux/compute_pass_encoder_handle.hpp
@@ -7,6 +7,7 @@ namespace systems::leal::campello_gpu {
     struct ComputePassEncoderHandle {
         VkCommandBuffer  commandBuffer;
         VkPipelineLayout pipelineLayout = VK_NULL_HANDLE; ///< Cached from last setPipeline call.
+        VkPipeline       boundPipeline  = VK_NULL_HANDLE; ///< Cached from last setPipeline call.
     };
 
 }