+++ src/vulkan_linux/compute_pipeline.cpp
@@ -13,10 +13,11 @@ ComputePipeline::~ComputePipeline() {
     auto handle = (ComputePipelineHandle *)native;
 
     vkDestroyPipeline(handle->device, handle->pipeline, nullptr);
-    vkDestroyPipelineLayout(handle->device, handle->pipelineLayout, nullptr);
+    if (handle->ownsPipelineLayout) {
+        vkDestroyPipelineLayout(handle->device, handle->pipelineLayout, nullptr);
+    }
 
     delete handle;
-    
 
 }