+++ src/vulkan_linux/compute_pipeline_handle.hpp
@@ -9,5 +9,6 @@ namespace systems::leal::campello_gpu
         VkDevice         device;
         VkPipeline       pipeline;
         VkPipelineLayout pipelineLayout;
+        bool             ownsPipelineLayout = false;
     };
 }