+++ src/vulkan_linux/render_pass_encoder_handle.hpp
@@ -7,6 +7,7 @@ namespace systems::leal::campello_gpu {
     struct RenderPassEncoderHandle {
         VkCommandBuffer  commandBuffer;
         VkPipelineLayout pipelineLayout   = VK_NULL_HANDLE; ///< Cached from last setPipeline call.
+        VkPipeline       boundPipeline    = VK_NULL_HANDLE; ///< Cached from last setPipeline call.
         // Swapchain path
         VkImage          currentSwapchainImage = VK_NULL_HANDLE;
         bool             isSwapchain           = false;